+++ spec/cucumber/formatter/json_spec.rb
@@ -715,6 +715,48 @@ module Cucumber
           end
         end
 
+        describe "with a scenario when only an around hook is failing" do
+          define_feature <<-FEATURE
+          Feature: Banana party
+
+            Scenario: Monkey eats bananas
+              Given there are bananas
+            FEATURE
+
+          define_steps do
+            Around() { |scenario, block| block.call; raise RuntimeError, "error" }
+            Given(/^there are bananas$/) {}
+          end
+
+          it "includes the around hook result in the json data" do
+            expect(load_normalised_json(@out)).to eq MultiJson.load(%{
+              [{"id": "banana-party",
+                "uri": "spec.feature",
+                "keyword": "Feature",
+                "name": "Banana party",
+                "line": 1,
+                "description": "",
+                "elements":
+                 [{"id": "banana-party;monkey-eats-bananas",
+                   "keyword": "Scenario",
+                   "name": "Monkey eats bananas",
+                   "line": 3,
+                   "description": "",
+                   "type": "scenario",
+                   "steps":
+                    [{"keyword": "Given ",
+                      "name": "there are bananas",
+                      "line": 4,
+                      "match": {"location": "spec/cucumber/formatter/json_spec.rb:728"},
+                      "result": {"status": "passed",
+                                 "duration": 1}}],
+                   "around":
+                    [{"match": {"location": "unknown_hook_location:1"},
+                      "result": {"status": "failed",
+                                 "error_message": "error (RuntimeError)\\n./spec/cucumber/formatter/json_spec.rb:727:in `Around'",
+                                 "duration": 1}}]}]}]})
+          end
+        end
       end
 
       def load_normalised_json(out)
@@ -727,7 +769,7 @@ module Cucumber
         json.each do |feature|
           elements = feature.fetch('elements') { [] }
           elements.each do |scenario|
-            ['steps', 'before', 'after'].each do |type|
+            ['steps', 'before', 'after', 'around'].each do |type|
               if scenario[type]
                 scenario[type].each do |step_or_hook|
                   normalise_json_step_or_hook(step_or_hook)